refactor: use freezed and json_serializable for  models and city_state

+++ weather_app/lib/screen/location_details/widget/aqi_forecast_widget.dart
@@ -78,7 +78,7 @@ class AQIForeCastWidget extends StatelessWidget {
                       children: airQualityModel.forecast.map((data) {
                         return AQIBarItem(
                           day: Utils().mapDayInWeek(data.day.weekday),
-                          aqi: data.avg,
+                          aqi: data.avg.toInt(),
                         );
                       }).toList(),
                     ),
